+++ app/src/main/res/xml/simply_baking_widget_info.xml
@@ -7,5 +7,5 @@
     android:previewImage="@drawable/example_appwidget_preview"
     android:resizeMode="horizontal|vertical"
     android:widgetCategory="home_screen"
-    android:updatePeriodMillis="3600000"
+    android:updatePeriodMillis="1000"
     android:autoAdvanceViewId="@id/widget_recipe_flipper"/>
